+++ spec/factories/deliveries.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :delivery do
+    delivery_type { 0 }
+    association :guild
+    association :notification
+  end
+end