+++ spec/models/guild_spec.rb
@@ -8,6 +8,8 @@ RSpec.describe Guild, type: :model do
   it { should have_many(:leader_guild_roles).class_name('GuildRole') }
   it { should have_many(:characters_with_leader_role).through(:leader_guild_roles).source(:character) }
   it { should have_many(:statics).dependent(:destroy) }
+  it { should have_many(:deliveries).dependent(:destroy) }
+  it { should have_many(:notifications).through(:deliveries) }
 
   it 'factory should be valid' do
     guild = build :guild