+++ spec/models/notification_spec.rb
@@ -1,4 +1,7 @@
 RSpec.describe Notification, type: :model do
+  it { should have_many(:deliveries).dependent(:destroy) }
+  it { should have_many(:guilds).through(:deliveries) }
+
   it 'factory should be valid' do
     notification = build :notification